Update to Game.java to include a "Free Kick"
I'm suggesting an update to Game.java to include an extra play type called "Free Kick".

Basically, when a team takes a safety, in the NCAA they're allowed to do a kickoff from a tee (in the NFL it's punt or placekick) but they must do it from the 20 yard line instead of the 35 (also true in the NFL), often resulting in a better starting position for the receiving team.

The play "Free kick" is basically identical to the kickoff except that the equation to determine where the offense will take over starts at 115 instead of 100 (to symbolize the 15 yards the ball is moved back in this type of kick).

I also changed the safety comment to say "Free kick" instead of kick off.

PS: This was the best way I could think to suggest a change without just sending you a reddit/github message. If you'd rather I didn't do this next time, please let me know.


Thanks!