TOD fix & adding time decrease to qbInterception
Made the turnover on downs section of runPlay an if/else so that runPlay doesn't keep going after changing possession for a turnover on downs. Otherwise the play keeps going but adds stats and logs any score/turnover as if the original offense is still on the field, but getEventPrefix logs as if the ball changed hands.

Added time decrease to qbInterception so that time actually runs off the clock when a pick happens. Put the time decrease after the log otherwise it looks like any logged play happening immediately following the interception took 0 seconds to occur.