Add news entries to log/add flavor text to bless/curse
Adds the blessed and cursed teams to the saved string and uses those to add news stories to week 0 about the team that was blessed/cursed.

Only one variant of flavor text right now: The cursed team will be hit by probation and recruiting restrictions, the blessed team is hiring a shot shot pro coach that is also an alum. I plan to expand on this.



Also, not sure if this was intentional or not, but only American Samoa was capable of being selected to be blessed. Removed that; I'll change it back if you meant for that.