Additional Week 0 Stories
Changed method of saving cursed and blessed teams. Team abbreviations are now saved and then findTeamAbbr is used to reconstruct a new Team object that is then used to add various info to the news story flavor texts.

Week 0 news stories are randomly picked from switch statements for both the blessed team (assuming there is one) and the cursed team. The default case prints no story right now (because it shouldn't be possible to trigger it). I may get cheeky and add a story to the default case about unforseen forces cursing/blessing the team just in case something unexplainable happens.

If there is no team blessed/cursed, their respective objects are set to null and no story is printed.