Added another new blessed team story ("sports enhancement drinks").
Also added a mechanism to allow for stories to be expanded upon in the middle of the season. In this case, a random week between 5 and 9 is chosen to have a news story posted to it, following up on the launch of the new drink. Depending on the poll position of the team, the story is different.