Improved Curse/Bless Week 0 Story Selector and Added New Stories
Changed the way curse/bless stories are picked, so that if a story with a bunch of potentially failable criteria is picked, and the criteria all fail, the switch statement is tried again with a new random case. If the same story is picked, repeat until a different one is picked.

Added a story about a team having a huge fight in the locker room that the coach failed to get wind of and act on fast enough. Added a follow up story to week 1 after the games have been played that checks to see how the QB and WR did, and posts one of four stories depending on QB completion percentage, WR1 (who is mentioned in the original story) targets vs WR2/WR3 targets, and whether or not the team won.