Uploading Progress for Home Desktop
Wrote skeleton and most of the mechanics of a story about a Conference TV Network or a Team TV Network.

Conf network is created if: There is not already a conf network, and a 20% chance passes.

Grants 5 prestige to each team in the conference upon creation.

Team TV Network is created if: There is not already a conf network or a team tv network and the team is this year's saveBless.

If Team TV Network exists, conf network only grants 1 prestige because you're already benefiting from being on TV for the most part.

If all these checks fail, go back and pick another story.