Adding Story About TV Network Deal for Conf or Team
IMPORTANT: If this story triggers, there's a 20% chance that if it hasn't already, a conference will be granted a TV deal, which gives every team a prestige bonus. I set the bonus to be 5, unless they had already gotten a team TV deal, in which case the bonus is 1. saveBless gets no additional prestige. If the conference already has a TV Deal, this story is skipped and a different one is printed in week 0.

I'd like your feedback on that for balance purposes or if we should just remove the prestige bonus all together.

---

Mechanics: If this story is triggered, a switch statement is performed to figure out what the name of the conference tv network will be. The total number of individual team tv deals that have been reached (aka the number of times this story has triggered to run the generic individual team tv network (think Longhorn Network) for teams within the conference), and their names are added to a string array.

If there are no team deals, and no conf deal, there's a 20% chance the conf tv network is formed.

If there are team deals and no conf deal, there's a 20% chance the conf tv network is formed and the story printed references having to come to an agreement with the teams that already had their own channel.

If there is no team deal for saveBless and no conf deal, and the 20% chance fails, saveBless gets their own network (special stories exist for HOL and ULA cause that's Hollywood and LA and well...TV is what they do)

Otherwise, the for loop is tried again for a different story.