Beginnings of restructuring of runPlay()
Coding checkpoint for work at home tonight or tomorrow.

Added 1st and 2nd half kickoffs, and added a half time to stop play after the end of the 2nd and reset possession to the away team.

Updated convGameTime() to handle the fact that during the course of a play, time may run past Q2 0:00.

Added checkGameSituation() based the discussed scenarios in #73.

Added a time decrement for huddling/time at the LOS based on checkGameSituation() if there is a running clock. Currently, because the rest of the plays have not been updated and the existing plays decrease time on their own, runningClock is never changed from false.